vietnamese-cyberbullying/requirements.txt
transformers
datasets
torch
pandas
grpcio
grpcio-tools
protobuf
tokenizers
numpy
